python/service_scripts/requirements.txt
psutil
tqdm
pyyaml
requests